Feature/reserva (#5)
* feat: crud de reservas criado

* fix: alterando tipagem de id do restaurante e usuario quando for fazer reserva

* add git ignore *.iml

* adição de classes de Test e IT para Avaliacao e mudando atributo nota para BigDecimal

* adição de classes de Test para caso de uso Avaliacao

* adição de classes de Test IT para Controlador de Avaliacao usando RestAssured

* feat: adicionado testes unitarios

* feat: adicionado testes bdd

---------

+++ src/main/java/br/com/fiap/reservarestaurante/application/usecases/avaliacao/create/AvaliacaoCreateUseCaseInput.java
@@ -1,3 +1,6 @@
 package br.com.fiap.reservarestaurante.application.usecases.avaliacao.create;
 
-public record AvaliacaoCreateUseCaseInput(String reservaId, String usuarioId, String comentario, float nota) {}
+import java.math.BigDecimal;
+
+public record AvaliacaoCreateUseCaseInput(
+    String reservaId, String usuarioId, String comentario, BigDecimal nota) {}
