+++ src/main/java/br/com/fiap/reservarestaurante/application/usecases/avaliacao/retrive/nota/NotaRestauranteGetByIdUseCaseOutput.java
@@ -1,14 +1,13 @@
 package br.com.fiap.reservarestaurante.application.usecases.avaliacao.retrive.nota;
 
 import br.com.fiap.reservarestaurante.application.domain.avaliacao.NotaRestaurante;
+import java.math.BigDecimal;
 
 public record NotaRestauranteGetByIdUseCaseOutput(
-        String restauranteId,
-        Long avaliacoes,
-        double nota) {
+    String restauranteId, Long avaliacoes, BigDecimal nota) {
 
-    public static NotaRestauranteGetByIdUseCaseOutput from(final NotaRestaurante notaRestaurante) {
-        return new NotaRestauranteGetByIdUseCaseOutput(notaRestaurante.restauranteId(), notaRestaurante.avaliacoes(), notaRestaurante.nota());
-    }
-
+  public static NotaRestauranteGetByIdUseCaseOutput from(final NotaRestaurante notaRestaurante) {
+    return new NotaRestauranteGetByIdUseCaseOutput(
+        notaRestaurante.restauranteId(), notaRestaurante.avaliacoes(), notaRestaurante.nota());
+  }
 }
